% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_f.R
\name{boxplot_f}
\alias{boxplot_f}
\title{Title}
\usage{
boxplot_f(df, var_name)
}
\arguments{
\item{df}{a dataframe (Given dataframe: Inpatient_Prospective_Payment)}

\item{var_name}{"Average.Total.Payments", "Average.Covered.Charges", "Average.Medicare.Payments"}
}
\value{
a boxplot by DRG codes
}
\description{
Title
}
