% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_f.R
\name{statistics_f}
\alias{statistics_f}
\title{Title}
\usage{
statistics_f(df, option)
}
\arguments{
\item{df}{a dataframe (Given dataframe: Inpatient_Prospective_Payment)}

\item{option}{"mean", "median", or "standard deviation" Have to choose one of them.}
}
\value{
the chosen statistic over all of the DRG codes for average Medicare payments
}
\description{
Title
}
